revert back to 64-bit numbers, delegate loop and function descent checks to parser

+++ EVIL.Lexical/TokenType.cs
@@ -58,7 +58,7 @@
 
         Identifier,
 
-        Decimal,
+        Double,
         Integer,
         HexInteger,
         String,
